requirements.txt
streamlit
requests
beautifulsoup4
nltk
torch
transformers
python-dotenv
pinecone-client
langchain
langchain-community
sentence-transformers
groq